[libgofono] Request initial properties for all objects
Even if properties are known at the moment when the object is created
(like the modems created as a result of receiving ModemAdded signal)
there's a chance that we miss PropertyChanged signal(s) from ofono
while we are creating the D-Bus proxy. To eliminate this possibility,
we need to always request current properties after registering the
PropertyChanged listener.